requirements.txt
pandas
xlrd
openpyxl
gradio
openai
ollama
python-dotenv